changing to standardize on "digits" being a u8

+++ src/worked_problems/euler_40.rs
@@ -7,7 +7,7 @@ d1 × d10 × d100 × d1000 × d10000 × d100000 × d1000000
 use crate::aux_funcs::{int_to_digits};
 
 pub fn euler40() -> u64 {
-    let mut out = 1;
+    let mut out = 1u64;
     let mut index = 10;
     let mut n = 2;
     let mut pos = 1;
@@ -18,7 +18,7 @@ pub fn euler40() -> u64 {
             // If we step past the index we need get the digits and go backward
             let mut digits = int_to_digits(n,10);
             digits.reverse();
-            out *= digits[pos-index];
+            out *= digits[pos-index] as u64;
             index *= 10;
             if index > 1000000 {
                 break
